+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# 0.5.0
+
+Add the `ExtendBlocks` trait and use it for `enable_bit_extending`. This allows
+using the method on arbitrary types.
+
+`ExtendBlocks` is implemented for:
+
+- `Box<[u32]>`
+- `Vec<u32>`
+- `SmallVec<[u32; N]>` (behind feature flag)
+
 # 0.4.0
 
 Remove `From<usize>` bound on `K` in `RawIndexMap`. It was unneeded, because